+++ boot.php
@@ -39,6 +39,7 @@ if (
 rex_api_function::register('massif_meta_get', Media\Api\GetMeta::class);
 rex_api_function::register('massif_image_get', Media\Api\GetImage::class);
 rex_api_function::register('upload_files', Form\Api\Upload::class);
+rex_api_function::register('massif_download', Form\Api\Download::class);
 
 
 // add own backend css