+++ lib/Form/Form.php
@@ -31,7 +31,7 @@ class Form
     $settings['template'] = 'email.contact-form';
     $settings['template_user'] = '';
     $settings['send_user_email'] = false;
-    $settings['skip_field_types'] = ['html', 'validate', 'action', 'csrf', 'mupload'];
+    $settings['skip_field_types'] = ['html', 'validate', 'action', 'csrf', 'mupload', 'multi_file_upload'];
     $settings['skip_fields'] = ['termsofuse_accepted', 'type'];
     $settings['replace_labels'] = ['subscribe_newsletter' => 'Newsletter', 'contact_city' => 'Stadt', 'confirm' => 'DSE akzeptiert?', 'context' => 'Interesse an', 'attachment' => 'Bewerbungsunterlagen', 'vacancy' => 'Stelle'];
     $settings['date_fields'] = ['dob'];